% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cases.R
\name{add_cases}
\alias{add_cases}
\title{Add new cases to Go.Data}
\usage{
add_cases(url, username, password, outbreak_id, data)
}
\arguments{
\item{url}{Insert the base URL for your instance of Go.Data here. Don't forget the forward slash "/" at end!}

\item{username}{The email address for your Go.Data login.}

\item{password}{The password for your Go.Data login}

\item{outbreak_id}{The id number for the outbreak for which you want to add data.}

\item{data}{The dataframe (as a list object) to import into Go.Data. See Details.}
}
\value{
Returns the status of the request.
}
\description{
A function to add new cases to a specific \code{outbreak_id}.
}
\details{
This function creates new records for cases. The data must be provided as a list object. Field names must match those used in Go.Data. You can use \code{get_data_model()} to determine appropriate field names.
}
\examples{
\dontrun{
url <- "https://MyGoDataServer.com/"
username <- "myemail@email.com"
password <- "mypassword"
outbreak_id <- "3b5554d7-2c19-41d0-b9af-475ad25a382b"

add_cases(url=url,
          username=username,
          password=password,
          outbreak_id=outbreak_id)
}
}
