% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_categories.R
\name{any_tokens}
\alias{any_tokens}
\title{Checks if there are any recognized tokens in the data provided}
\usage{
any_tokens(data, language_tokens)
}
\arguments{
\item{data}{A data frame (or data frame extension)}

\item{language_tokens}{A data frame (or data frame extension) containing
the translations. Output from \code{get_language_tokens()}}
}
\value{
Boolean logical (TRUE or FALSE)
}
\description{
Checks if there are any recognized tokens in the data provided
}
\keyword{internal}
