% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_downloader.R
\name{batch_downloader}
\alias{batch_downloader}
\title{Function to manage batch downloads}
\usage{
batch_downloader(
  url = url,
  username = username,
  password = password,
  api_call_n = api_call_n,
  api_call_get = api_call_get,
  batch_size = batch_size
)
}
\arguments{
\item{url}{Insert the base URL for your instance of Go.Data here. Don't forget the forward slash "/" at end!}

\item{username}{The email address for your Go.Data login.}

\item{password}{The password for your Go.Data login}

\item{api_call_n}{The API url to get the number of records.}

\item{api_call_get}{The API url to GET the records.}

\item{batch_size}{Specifies the number of records to retrieve in each iteration.}
}
\value{
Returns a data frame. Some fields, such as addresses, hospitalization history, and questionnaire fields will require further unnesting. See the \code{\link[tidyr]{unnest}} function.
}
\description{
A housekeeping function to do batch downloads.
}
\examples{
\dontrun{
url <- "https://MyGoDataServer.com/"
username <- "myemail@email.com"
password <- "mypassword"
outbreak_id <- "3b5554d7-2c19-41d0-b9af-475ad25a382b"

cases <- get_cases(url=url,
                   username=username,
                   password=password,
                   outbreak_id=outbreak_id)
}
}
