% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases_from_contacts.R
\name{cases_from_contacts}
\alias{cases_from_contacts}
\title{Pull out all cases that used to be contacts}
\usage{
cases_from_contacts(cases_clean)
}
\arguments{
\item{cases_clean}{The cleaned case data. Case data is returned by
\code{\link[=get_cases]{get_cases()}} and cleaned by \code{\link[=clean_cases]{clean_cases()}}.}
}
\value{
A tibble containing the cases that used to be contacts.
}
\description{
Pull out all cases that used to be contacts
}
\examples{
\dontrun{
url <- "https://MyGoDataServer.com/"
username <- "myemail@email.com"
password <- "mypassword"
outbreak_id <- "3b5554d7-2c19-41d0-b9af-475ad25a382b"

cases <- get_cases(
  url = url,
  username = username,
  password = password,
  outbreak_id = outbreak_id
)

# other cleaned data required for `clean_cases()`
cases_vacc_history_clean <- clean_case_vax_history(cases = cases)
cases_address_history_clean <- clean_case_address_history(cases = cases)
cases_dateranges_history_clean <- clean_case_med_history(cases = cases)

cases_clean <- clean_cases(
  cases = cases,
  cases_address_history_clean = cases_address_history_clean,
  cases_vacc_history_clean = cases_vacc_history_clean,
  cases_dateranges_history_clean = cases_dateranges_history_clean
)

cases_from_contacts <- cases_from_contacts(cases_clean = cases_clean)
}
}
