% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_values_to_tokens.R
\name{change_values_to_tokens}
\alias{change_values_to_tokens}
\title{Replace column values with Go.Data tokens}
\usage{
change_values_to_tokens(data, column_name, language_tokens = language_tokens)
}
\arguments{
\item{data}{A data frame to be imported into Go.Data}

\item{column_name}{The name of the column to be replaced with tokens.}

\item{language_tokens}{A language token object by using the \code{get_language_tokens()} function.}
}
\value{
A data frame
}
\description{
A function to replace values in a field with tokens from reference data
}
\details{
A useful function when preparing data to be imported into Go.Data.
}
\examples{
\dontrun{
url <- "https://MyGoDataServer.com/"
username <- "myemail@email.com"
password <- "mypassword"
outbreak_id <- "3b5554d7-2c19-41d0-b9af-475ad25a382b"

language_tokens <- get_language_tokens(url=url,
                                       username=username,
                                       password=password,
                                       language="english_us")

#Example data frame to be imported into Go.Data
x <- data.frame(visualId=c("CASE-001","CASE-002"),
                gender=c("Male","Female"))
x_new <- change_values_to_tokens(data=x, column_name="gender", language_tokens=language_tokens)
}
}
