% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_godata_url.R
\name{check_godata_url}
\alias{check_godata_url}
\title{Check if the provided URL is valid}
\usage{
check_godata_url(url = url)
}
\arguments{
\item{url}{Insert the base URL for your instance of Go.Data here. Don't forget the forward slash "/" at end!}
}
\value{
boolean
}
\description{
Check if the provided URL is valid
}
\examples{
\dontrun{
url <- "https://MyGoDataServer.com/"
check_godata_url(url=url)
}
}
