% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_case_address_history.R
\name{clean_case_address_history}
\alias{clean_case_address_history}
\title{Extract address information from case data}
\usage{
clean_case_address_history(cases)
}
\arguments{
\item{cases}{A tibble with case data. Case data is returned by
\code{\link[=get_cases]{get_cases()}}.}
}
\value{
A tibble with address information from cases data.
}
\description{
This function un-nests and cleans the address data and stores
it in a standalone table with all addresses, even if there is more than 1
per person.
}
\examples{
\dontrun{
url <- "https://MyGoDataServer.com/"
username <- "myemail@email.com"
password <- "mypassword"
outbreak_id <- "3b5554d7-2c19-41d0-b9af-475ad25a382b"

cases <- get_cases(
  url = url,
  username = username,
  password = password,
  outbreak_id = outbreak_id
)

case_address_history <- clean_case_address_history(cases = cases)
}
}
