% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_case_med_history.R
\name{clean_case_med_history}
\alias{clean_case_med_history}
\title{Extracts and cleans medical history from case data}
\usage{
clean_case_med_history(cases)
}
\arguments{
\item{cases}{A tibble with case data. Case data is returned by
\code{\link[=get_cases]{get_cases()}}.}
}
\value{
A tibble with information on isolation and hospitalization history.
}
\description{
This function un-nests and cleans date ranges of isolation and
hospitalization history and stores it in a standalone table.
}
\examples{
\dontrun{
url <- "https://MyGoDataServer.com/"
username <- "myemail@email.com"
password <- "mypassword"
outbreak_id <- "3b5554d7-2c19-41d0-b9af-475ad25a382b"

cases <- get_cases(
  url = url,
  username = username,
  password = password,
  outbreak_id = outbreak_id
)

cases_med_history <- clean_case_med_history(cases = cases)
}
}
