% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_case_vax_history.R
\name{clean_case_vax_history}
\alias{clean_case_vax_history}
\title{Cleans vaccination data from case data}
\usage{
clean_case_vax_history(cases, language_tokens)
}
\arguments{
\item{cases}{A tibble with address information from cases data.}

\item{language_tokens}{A tibble of language tokens returned by
\code{\link[=get_language_tokens]{get_language_tokens()}} to translate the string tokens in the data.}
}
\value{
A tibble with cleaned and un-nested vaccination history data.
}
\description{
Cleans and un-nests vaccination history, where vaccination is
complete, from case data. Case data is returned from \code{\link[=get_cases]{get_cases()}}.
}
\examples{
#' \dontrun{
url <- "https://MyGoDataServer.com/"
username <- "myemail@email.com"
password <- "mypassword"
outbreak_id <- "3b5554d7-2c19-41d0-b9af-475ad25a382b"

cases <- get_cases(
  url = url,
  username = username,
  password = password,
  outbreak_id = outbreak_id
)

language_tokens <- get_language_tokens(
  url = url,
  username = username,
  password = password,
  language = "english_us"
)

vax_history <- clean_case_vax_history(
  cases = cases,
  language_tokens = language_tokens
)
}
}
