% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_contact_vax_history.R
\name{clean_contact_vax_history}
\alias{clean_contact_vax_history}
\title{Cleans vaccination data from contact data}
\usage{
clean_contact_vax_history(contacts)
}
\arguments{
\item{contacts}{A tibble with address information from contact data.}
}
\value{
A tibble with cleaned and un-nested vaccination history data.
}
\description{
Cleans and un-nests vaccination history, where vaccination is
complete, from contact data. Contact data is returned from
\code{\link[=get_contacts]{get_contacts()}}.
}
\examples{
\dontrun{
url <- "https://MyGoDataServer.com/"
username <- "myemail@email.com"
password <- "mypassword"
outbreak_id <- "3b5554d7-2c19-41d0-b9af-475ad25a382b"

contacts <- get_contacts(
  url = url,
  username = username,
  password = password,
  outbreak_id = outbreak_id
)

vax_history <- clean_contact_vax_history(contacts = contacts)
}
}
