% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_contacts_of_contacts.R
\name{clean_contacts_of_contacts}
\alias{clean_contacts_of_contacts}
\title{Clean contacts of contacts data}
\usage{
clean_contacts_of_contacts(
  contacts_of_contacts,
  contacts_of_contacts_address_history_clean,
  contacts_of_contacts_vacc_history_clean
)
}
\arguments{
\item{contacts_of_contacts}{A \code{tibble} containing the contacts of contacts
data.}

\item{contacts_of_contacts_address_history_clean}{A \code{tibble} containing the
cleaned address history data from contacts of contacts (data is cleaned by
\code{\link[=clean_contacts_of_contacts_address_history]{clean_contacts_of_contacts_address_history()}}).}

\item{contacts_of_contacts_vacc_history_clean}{A \code{tibble} containing the
cleaned vaccination history from contacts of contacts (data is cleaned by
\code{\link[=clean_contacts_of_contacts_vax_history]{clean_contacts_of_contacts_vax_history()}}).}
}
\value{
A \code{tibble} containing the cleaned contacts of contacts data.
}
\description{
Cleans and un-nests contacts of contacts data. Contacts of
contacts data is returned by \code{\link[=get_contacts_of_contacts]{get_contacts_of_contacts()}}.
}
\examples{
\dontrun{
  url <- "https://MyGoDataServer.com/"
  username <- "myemail@email.com"
  password <- "mypassword"
  outbreak_id <- "3b5554d7-2c19-41d0-b9af-475ad25a382b"

contacts_of_contacts <- get_contacts_of_contacts(
  url = url,
  username = username,
  password = password,
  outbreak_id = outbreak_id
)

locations <- get_locations(
  url = url,
  username = username,
  password = password
)

locations_clean <- clean_locations(locations = locations)

contacts_of_contacts_address_history_clean <- clean_contacts_of_contacts_address_history(
  contacts_of_contacts = contacts_of_contacts,
  locations_clean = locations_clean
)

contacts_of_contacts_vacc_history_clean <- clean_contacts_of_contacts_vax_history(
  contacts_of_contacts = contacts_of_contacts
)

contacts_of_contacts_clean <- clean_contacts_of_contacts(
  contacts_of_contacts = contacts_of_contacts,
  contacts_of_contacts_address_history_clean = contacts_of_contacts_address_history_clean,
  contacts_of_contacts_vacc_history_clean = contacts_of_contacts_vacc_history_clean
)
}
}
