% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_contacts_of_contacts_address_history.R
\name{clean_contacts_of_contacts_address_history}
\alias{clean_contacts_of_contacts_address_history}
\title{Extracts address information from contacts of contacts data}
\usage{
clean_contacts_of_contacts_address_history(
  contacts_of_contacts,
  locations_clean
)
}
\arguments{
\item{contacts_of_contacts}{A\code{tibble} with contacts of contacts data.
Contacts of contacts data is returned by \code{\link[=get_contacts_of_contacts]{get_contacts_of_contacts()}}.}

\item{locations_clean}{A \code{tibble} with cleaned location data. Location data
is returned by \code{\link[=get_locations]{get_locations()}} and cleaned by \code{\link[=clean_locations]{clean_locations()}}.}
}
\value{
A \code{tibble} with address information from contacts of contacts data.
}
\description{
This function un-nests and cleans the address data and stores
it in a standalone table with all addresses, even if there is more than 1
per person.
}
\examples{
\dontrun{
url <- "https://MyGoDataServer.com/"
username <- "myemail@email.com"
password <- "mypassword"
outbreak_id <- "3b5554d7-2c19-41d0-b9af-475ad25a382b"

contacts_of_contacts <- get_contacts_of_contacts(
  url = url,
  username = username,
  password = password,
  outbreak_id = outbreak_id
)

locations <- get_locations(
  url = url,
  username = username,
  password = password
)
locations_clean <- clean_locations(locations = locations)

contact_of_contacts_address_history <- clean_contacts_of_contacts_address_history(
  contacts_of_contacts = contacts_of_contacts,
  locations_clean = locations_clean
)
}
}
