% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_events.R
\name{clean_events}
\alias{clean_events}
\title{Clean events data}
\usage{
clean_events(events, locations_clean)
}
\arguments{
\item{events}{A \code{tibble} with events data. Events data is returned by
\code{\link[=get_events]{get_events()}}.}

\item{locations_clean}{A \code{tibble} with cleaned location data. Location data
is returned by \code{\link[=get_locations]{get_locations()}} and cleaned by \code{\link[=clean_locations]{clean_locations()}}.
Make sure the locations data is cleaned prior to supplying it to
\code{clean_events()}.}
}
\value{
A \code{tibble} with cleaned events data.
}
\description{
Cleans and un-nests events data which is returned from
\code{\link[=get_events]{get_events()}}.
}
\examples{
\dontrun{
url <- "https://MyGoDataServer.com/"
username <- "myemail@email.com"
password <- "mypassword"
outbreak_id <- "3b5554d7-2c19-41d0-b9af-475ad25a382b"

events <- get_events(
  url = url,
  username = username,
  password = password,
  outbreak_id = outbreak_id
)

locations <- get_locations(
  url = url,
  username = username,
  password = password
)
locations_clean <- clean_locations(locations = locations)

clean_events <- clean_events(
  events = events,
  locations_clean = locations_clean)
}
}
