% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_followups.R
\name{clean_followups}
\alias{clean_followups}
\title{Clean followup data}
\usage{
clean_followups(followups, contacts_address_history_clean, language_tokens)
}
\arguments{
\item{followups}{A \code{tibble} with events data. Followup data is returned by
\code{\link[=get_followups]{get_followups()}}.}

\item{contacts_address_history_clean}{A \code{tibble} with cleaned address
history data from contacts. Contacts data is returned by \code{\link[=get_contacts]{get_contacts()}}
and cleaned by \code{\link[=clean_contact_address_history]{clean_contact_address_history()}}.}
}
\value{
A \code{tibble} with cleaned followup data.
}
\description{
Cleans and un-nests followup data which is returned from
\code{\link[=get_followups]{get_followups()}}
}
\examples{
\dontrun{
url <- "https://MyGoDataServer.com/"
username <- "myemail@email.com"
password <- "mypassword"
outbreak_id <- "3b5554d7-2c19-41d0-b9af-475ad25a382b"

followups <- get_followups(
  url = url,
  username = username,
  password = password,
  outbreak_id = outbreak_id
)

contacts <- get_contacts(
  url = url,
  username = username,
  password = password,
  outbreak_id = outbreak_id
)

locations <- get_locations(
  url = url,
  username = username,
  password = password
)

locations_clean <- clean_locations(locations = locations)

language_tokens <- get_language_tokens(
  url = url,
  username = username,
  password = password,
  language = "english_us"
)

contacts_address_history_clean <- clean_contact_address_history(
  contacts = contacts,
  locations_clean = locations_clean,
  language_tokens = language_tokens
)

followups_clean <- clean_followups(
  followups = followups,
  contacts_address_history_clean = contacts_address_history_clean,
  language_tokens = language_tokens
)
}
}
