% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_locations.R
\name{clean_locations}
\alias{clean_locations}
\title{Cleans location data}
\usage{
clean_locations(locations)
}
\arguments{
\item{locations}{A \code{\link{tibble}} containing locations data. This is the data
returned from \code{\link[=get_locations]{get_locations()}}}
}
\value{
A \code{tibble} containing the cleaned and rearranged location data.
}
\description{
Rearrange via joins to get into more usable hierarchy format,
these can then be joined to cases, contacts, etc for further analysis
}
\examples{
\dontrun{
url <- "https://MyGoDataServer.com/"
username <- "myemail@email.com"
password <- "mypassword"

locations <- get_locations(
  url = url,
  username = username,
  password = password
)

clean_locations(locations = locations)
}
}
