% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_relationships.R
\name{clean_relationships}
\alias{clean_relationships}
\title{Cleans relationship data}
\usage{
clean_relationships(relationships, language_tokens)
}
\arguments{
\item{relationships}{A \code{tibble} of relationship data. Relationship data is
returned by \code{\link[=get_relationships]{get_relationships()}}.}
}
\value{
A \code{tibble} with clean relationship data.
}
\description{
Cleans and un-nests relationship data. Relationship data is
returned by \code{\link[=get_relationships]{get_relationships()}}.
}
\examples{
\dontrun{
url <- "https://MyGoDataServer.com/"
username <- "myemail@email.com"
password <- "mypassword"
outbreak_id <- "3b5554d7-2c19-41d0-b9af-475ad25a382b"

relationships <- get_relationships(
  url = url,
  username = username,
  password = password,
  outbreak_id = outbreak_id
)

language_tokens <- get_language_tokens(
  url = url,
  username = username,
  password = password,
  language = "english_us"
)

clean_relationships <- clean_relationships(
  relationships,
  language_tokens = language_tokens
)
}
}
