% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_teams.R
\name{clean_teams}
\alias{clean_teams}
\title{Clean teams data}
\usage{
clean_teams(teams)
}
\arguments{
\item{teams}{A \code{tibble} containing teams data. Teams data is returned by
\code{\link[=get_teams]{get_teams()}}.}
}
\value{
A \code{tibble} of cleaned teams data
}
\description{
Cleans and un-nests teams data. Teams data is returned by
\code{\link[=get_teams]{get_teams()}}.
}
\examples{
\dontrun{
url <- "https://MyGoDataServer.com/"
username <- "myemail@email.com"
password <- "mypassword"

teams <- get_teams(
  url = url,
  username = username,
  password = password
)

clean_teams <- clean_teams(teams)
}
}
