% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_users.R
\name{clean_users}
\alias{clean_users}
\title{Cleans users data}
\usage{
clean_users(users, language_tokens)
}
\arguments{
\item{users}{A \code{tibble} containing users data. Users data is returned by
\code{\link[=get_users]{get_users()}}.}
}
\value{
A \code{tibble} with cleaned users data.
}
\description{
Cleans and un-nests users data. Users data is returned by
\code{\link[=get_users]{get_users()}}.
}
\examples{
\dontrun{
url <- "https://MyGoDataServer.com/"
username <- "myemail@email.com"
password <- "mypassword"

users <- get_users(
  url = url,
  username = username,
  password = password
)

language_tokens <- get_language_tokens(
  url = url,
  username = username,
  password = password,
  language = "english_us"
)

clean_users <- clean_users(users = users, language_tokens = language_tokens)
}
}
