% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contacts_per_case.R
\name{contacts_per_case}
\alias{contacts_per_case}
\title{Counts the number of contacts per case from relationship data}
\usage{
contacts_per_case(relationships_clean)
}
\arguments{
\item{relationships_clean}{A \code{tibble} with the cleaned relationship data.
Relationship data is returned by \code{\link[=get_relationships]{get_relationships()}} and cleaned by
\code{\link[=clean_relationships]{clean_relationships()}}.}
}
\value{
A \code{tibble} with the number of contacts associated to each source
person
}
\description{
Uses cleaned relationship data to tally the number of contacts
per case. Relationship data is returned by \code{\link[=get_relationships]{get_relationships()}} and
cleaned by \code{\link[=clean_relationships]{clean_relationships()}}.
}
\examples{
\dontrun{
url <- "https://MyGoDataServer.com/"
username <- "myemail@email.com"
password <- "mypassword"
outbreak_id <- "3b5554d7-2c19-41d0-b9af-475ad25a382b"

relationships <- get_relationships(
  url = url,
  username = username,
  password = password,
  outbreak_id = outbreak_id
)

clean_relationships <- clean_relationships(relationships)

contacts_per_case <- contacts_per_case(clean_relationships)
}
}
