% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_location_tree.R
\name{expand_location_tree}
\alias{expand_location_tree}
\title{Expand the tree for location data from Go.Data}
\usage{
expand_location_tree(locations = locations)
}
\arguments{
\item{locations}{A data frame of locations. Can be retrieved using get_locations().}
}
\value{
Returns a data frame of locations associated with Go.Data instance including all location id numbers at different administrative levels.
}
\description{
A function to add location id numbers at all
administrative levels.
}
\examples{
\dontrun{
url <- "https://MyGoDataServer.com/"
username <- "myemail@email.com"
password <- "mypassword"

locations <- get_locations(url=url,
                           username=username,
                           password=password)

locations_clean <- expand_location_tree(locations)

#All in one step
locations_clean <- expand_location_tree(get_locations(url,username,password))
}
}
