% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_outbreaks.R
\name{get_all_outbreaks}
\alias{get_all_outbreaks}
\title{Get a list of all outbreaks and their attributes}
\usage{
get_all_outbreaks(url = url, username = username, password = password)
}
\arguments{
\item{url}{Insert the base URL for your instance of Go.Data here. Don't forget the forward slash "/" at end!}

\item{username}{The email address for your Go.Data login.}

\item{password}{The password for your Go.Data login}
}
\value{
Returns data frame of outbreaks. The resulting list is filtered by the user's permissions: only outbreaks for which the user has access will be returned.
}
\description{
Get a list of all outbreaks and their attributes
}
\examples{
\dontrun{
url <- "https://MyGoDataServer.com/"
username <- "myemail@email.com"
password <- "mypassword"

outbreaks <- get_all_outbreaks(url=url, username=username, password=password)
}
}
