% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contacts_of_contacts.R
\name{get_contacts_of_contacts}
\alias{get_contacts_of_contacts}
\title{Download contacts-of-contacts from Go.Data (version 2.38.0 or earlier)}
\usage{
get_contacts_of_contacts(
  url = url,
  username = username,
  password = password,
  outbreak_id = outbreak_id,
  batch_size = 50000
)
}
\arguments{
\item{url}{Insert the base URL for your instance of Go.Data here. Don't forget the forward slash "/" at end!}

\item{username}{The email address for your Go.Data login.}

\item{password}{The password for your Go.Data login}

\item{outbreak_id}{The id number for the outbreak for which you want to download contacts-of-contacts.}

\item{batch_size}{For large datasets, specify the number of records to retrieve in each iteration.}
}
\value{
Returns a data frame of contacts-of-contacts. Some fields, such as addresses, hospitalization history, and questionnaire fields will require further unnesting. See the \code{tidyr::unnest()} function.
}
\description{
A function to retrieve the case data for a
specific \code{outbreak_id}. This function relies
on the \verb{/outbreaks/\{id\}/contacts-of-contacts}
API endpoint. Records are imported in
iterative batches and then appended together
into a single data frame.
}
\details{
This function works on all versions of Go.Data.
However, if you are using version 2.38.1 or
greater, then we recommend using \code{\link{get_contacts_of_contacts2}},
which has additional features and improved
performance.
}
\examples{
\dontrun{
url <- "https://MyGoDataServer.com/"
username <- "myemail@email.com"
password <- "mypassword"
outbreak_id <- "3b5554d7-2c19-41d0-b9af-475ad25a382b"

contacts_of_contacts <- get_contacts_of_contacts(url=url,
                                                 username=username,
                                                 password=password,
                                                 outbreak_id=outbreak_id)
}
}
