% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_model.R
\name{get_data_model}
\alias{get_data_model}
\title{Get a data model from Go.Data}
\usage{
get_data_model(
  url = url,
  username = username,
  password = password,
  model = c("case", "contact", "followUp", "address", "age", "relationship", "event",
    "dateRange", "contactOfContact")
)
}
\arguments{
\item{url}{Insert the base URL for your instance of Go.Data here. Don't forget the forward slash "/" at end!}

\item{username}{The email address for your Go.Data login.}

\item{password}{The password for your Go.Data login}

\item{model}{The data model to be retrieved.}
}
\value{
Returns a data frame. Data frame will have 0 rows, but its columns and column names can be used as a template.
}
\description{
A function to retrieve the data model (data template) for a specific type of object. This is a helper function to make sure that other get_* functions do not return empty results.
}
\examples{
\dontrun{
url <- "https://MyGoDataServer.com/"
username <- "myemail@email.com"
password <- "mypassword"

case_data_model <- get_data_model(url=url,
                                  username=username,
                                  password=password,
                                  model="case")
}
}
