% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_date_range.R
\name{get_date_range}
\alias{get_date_range}
\title{Get minimum and maximum date for a vector of dates}
\usage{
get_date_range(dates)
}
\arguments{
\item{dates}{character vector of dates to extract the date range from}
}
\value{
Returns a list object with two values (minimum and maximum date)
}
\description{
This function will attempt to auto-detect the date format using lubridate,
then determine the minimum and maximum for the submitted vector of dates.
It is intended as a helper function for selecting date ranges and adding
them to queries.
}
\examples{
# Create a character vector of dates:
x <- c("2022-07-15", "2021-08-09", NA, "2022-08-03")

# Get date range:
daterange <- get_date_range(dates = x)

# View the result:
daterange
}
\author{
Amy Mikhail, \email{amy.mikhail@gmail.com}
}
