% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_events.R
\name{get_events}
\alias{get_events}
\title{Download events from Go.Data (version 2.38.0 or earlier)}
\usage{
get_events(
  url = url,
  username = username,
  password = password,
  outbreak_id = outbreak_id,
  batch_size = 50000
)
}
\arguments{
\item{url}{Insert the base URL for your instance of Go.Data here. Don't forget the forward slash "/" at end!}

\item{username}{The email address for your Go.Data login.}

\item{password}{The password for your Go.Data login}

\item{outbreak_id}{The id number for the outbreak for which you want to download events.}

\item{batch_size}{For large datasets, specify the number of records to retrieve in each iteration.}
}
\value{
Returns data frame of events. may require further unnesting. See the \code{tidyr::unnest()} function.
}
\description{
A function to retrieve the event data for a
specific \code{outbreak_id}. This function relies
on the \verb{/outbreaks/\{id\}/events} API endpoint.
Records are imported in iterative batches
and then appended together into a single data
frame.
}
\details{
This function works on all versions of Go.Data.
However, if you are using version 2.38.1 or
greater, then we recommend using \code{\link{get_events2}},
which has additional features and improved
performance.
}
\examples{
\dontrun{
url <- "https://MyGoDataServer.com/"
username <- "myemail@email.com"
password <- "mypassword"
outbreak_id <- "3b5554d7-2c19-41d0-b9af-475ad25a382b"

events <- get_events(url=url,
                     username=username,
                     password=password,
                     outbreak_id=outbreak_id)
}
}
