% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_followups.R
\name{get_followups}
\alias{get_followups}
\title{Download follow-up of contacts from Go.Data (version 2.38.0 or earlier)}
\usage{
get_followups(
  url = url,
  username = username,
  password = password,
  outbreak_id = outbreak_id,
  batch_size = 50000
)
}
\arguments{
\item{url}{Insert the base URL for your instance of Go.Data here. Don't forget the forward slash "/" at end!}

\item{username}{The email address for your Go.Data login.}

\item{password}{The password for your Go.Data login}

\item{outbreak_id}{The id number for the outbreak for which you want to download follow-ups.}

\item{batch_size}{For large datasets, specify the number of records to retrieve in each iteration.}
}
\value{
Returns data frame of follow-ups. Some fields may require further unnesting. See the tidyr::unnest() function.
}
\description{
Download follow-up of contacts from Go.Data (version 2.38.0 or earlier)
}
\examples{
url <- "https://MyGoDataServer.com/"
username <- "myemail@email.com"
password <- "mypassword"
outbreak_id <- "3b5554d7-2c19-41d0-b9af-475ad25a382b"

followups <- get_followups(url=url, username=username, password=password, outbreak_id=outbreak_id)
}
