% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lab_fields.R
\name{get_lab_fields}
\alias{get_lab_fields}
\title{Get lookup table of Go.Data lab result fields}
\usage{
get_lab_fields(url = url, username = username, password = password)
}
\arguments{
\item{url}{URL (web address) for Go.Data instance}

\item{username}{User email address used to log in to Go.Data}

\item{password}{User password used to log in to Go.Data}
}
\value{
Returns a list of lab fields including values to map where relevant
}
\description{
This function creates a look-up table of lab result fields (both core and
user-designed questionnaire) and their hard-coded values, where relevant.
The fields and values are extracted for the active outbreak of a user's
Go.Data instance.  Values will include user-defined values that were added
to the reference sets in Go.Data.  The fields and values can then be mapped
to a laboratory data file, in preparation for importing to Go.Data.
}
\examples{
\dontrun{
# Get lab fields from a specific Go.Data instance:
labvars <- get_lab_fields(url = url, 
                          username = username, 
                          password = password)

# View the result:
labvars
}
}
\author{
Amy Mikhail, \email{amy.mikhail@gmail.com}
}
