% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_languages.R
\name{get_languages}
\alias{get_languages}
\title{Get lanuages in Go.Data}
\usage{
get_languages(url, username, password)
}
\arguments{
\item{url}{Insert the base URL for your instance of Go.Data here. Don't
forget the forward slash "/" at end!}

\item{username}{The email address for your Go.Data login.}

\item{password}{The password for your Go.Data login}
}
\value{
Returns data frame of languages available in Go.Data.
}
\description{
Get lanuages in Go.Data
}
\examples{
\dontrun{
url <- "https://MyGoDataServer.com/"
username <- "myemail@email.com"
password <- "mypassword"
language <- "english_us"

languages <- get_languages(
  url = url,
  username = username,
  password = password
)
}
}
