% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_locations.R
\name{get_locations}
\alias{get_locations}
\title{Get location data from Go.Data}
\usage{
get_locations(url = url, username = username, password = password)
}
\arguments{
\item{url}{Insert the base URL for your instance of Go.Data here. Don't
forget the forward slash "/" at end!}

\item{username}{The email address for your Go.Data login.}

\item{password}{The password for your Go.Data login}
}
\value{
Returns a data frame of locations associated with Go.Data instance.
}
\description{
A function to retreive a list of all locations
and their attributes, across all outbreaks on
the instance (since this is a system-level API
endpoint). This function relies on the \verb{\\locations}
API endpoint.
}
\examples{
\dontrun{
url <- "https://MyGoDataServer.com/"
username <- "myemail@email.com"
password <- "mypassword"

locations <- get_locations(
  url = url,
  username = username,
  password = password
)
}
}
