% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference_data.R
\name{get_reference_data}
\alias{get_reference_data}
\title{Get reference data from Go.Data}
\usage{
get_reference_data(url, username, password)
}
\arguments{
\item{url}{Insert the base URL for your instance of Go.Data here. Don't
forget the forward slash "/" at end!}

\item{username}{The email address for your Go.Data login.}

\item{password}{The password for your Go.Data login}
}
\value{
Returns data frame of reference data associated with Go.Data instance.
}
\description{
A function to retreive a list of all reference
data and their attributes, across all outbreaks
on the instance (since this is a system-level API
endpoint). This function relies on the
\verb{\\reference-data} API endpoint.
}
\examples{
\dontrun{
url <- "https://MyGoDataServer.com/"
username <- "myemail@email.com"
password <- "mypassword"

reference_data <- get_reference_data(
  url = url,
  username = username,
  password = password
)
}
}
