% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_teams.R
\name{get_teams}
\alias{get_teams}
\title{Get a list of all teams and their attributes, across all outbreaks on the instance (since this is a system-level API endpoint).}
\usage{
get_teams(url = url, username = username, password = password)
}
\arguments{
\item{url}{Insert the base URL for your instance of Go.Data here. Don't forget the forward slash "/" at end!}

\item{username}{The email address for your Go.Data login.}

\item{password}{The password for your Go.Data login}
}
\value{
Returns data frame of teams associated with Go.Data instance.
}
\description{
Get a list of all teams and their attributes, across all outbreaks on the instance (since this is a system-level API endpoint).
}
\examples{
\dontrun{
url <- "https://MyGoDataServer.com/"
username <- "myemail@email.com"
password <- "mypassword"

teams <- get_teams(url=url, username=username, password=password)
}
}
