% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab2godata_wrapper.R
\name{lab2godata_wrapper}
\alias{lab2godata_wrapper}
\title{Wrapper function to link lab data to Go.Data cases and lab records}
\usage{
lab2godata_wrapper(
  url = "http://localhost:3000/",
  username,
  password,
  outbreak = "active",
  reason = c("link new", "edit lab", "add sequencing"),
  datequery = "epibuffer",
  daterangeformat = c("ymd", "dmy", "mdy"),
  epiwindow,
  method = c("fuzzy", "exact"),
  matchcols = c("names & dob", "names & age", "names", "doc ID"),
  labdata,
  basedatecol,
  firstnamecol = NULL,
  lastnamecol = NULL,
  dobcol = NULL,
  agecol = NULL,
  docidcol = NULL
)
}
\arguments{
\item{url}{The URL or web address for your Go.Data instance}

\item{username}{Your Go.Data username (email address used for logging in)}

\item{password}{Your Go.Data password}

\item{outbreak}{Default is your active outbreak; else enter the outbreak ID}

\item{reason}{Match reason: one of "link new", "edit lab" or "add sequencing"}

\item{datequery}{Default is "epibuffer"; else choose from "date range",
"epiwindow", or "sampledates"}

\item{daterangeformat}{One of "ymd", "dmy" or "mdy" (y=year, m=month, d=day)}

\item{epiwindow}{limit in days between dates from labdata and lookuptable}

\item{method}{Method to match on; one of "fuzzy" or "exact"}

\item{matchcols}{One of "names & dob", "names & age", "names", or "doc ID"}

\item{labdata}{Lab data for matching; either data.frame or data.table}

\item{basedatecol}{Name of lab column containing sample dates (mandatory)}

\item{firstnamecol}{Name of lab column containing first names, if needed}

\item{lastnamecol}{Name of lab column containing last names, if needed}

\item{dobcol}{Name of lab column containing birthdates, if needed}

\item{agecol}{Name of lab column containing age in years, if needed}

\item{docidcol}{Name of lab column containing document ID, if needed}
}
\value{
matched lab results in \code{match_report} and \code{matched_data}
}
\description{
This function is a wrapper function that imports lab data, matches it to
existing Go.Data case or lab records by patient identifier columns and
returns the updated lab records with the match IDs appended.  It is intented
to help Go.Data users link lab records to Go.Data case or lab record IDs by
exact or fuzzy matching on patient demographics, as this facility is not yet
available within Go.Data. The exported output file may be directly imported
to Go.Data using the bulk import facility.  Columns within the file can be
mapped to Go.Data fields in the normal way;
see \href{https://apps.who.int/iris/bitstream/handle/10665/332255/WHO-2019-nCoV-Go.data_manual-2020.2-eng.pdf?sequence=1&isAllowed=y}{Go.Data user manual chapter nine})
}
\examples{
\dontrun{
# Set file path of lab data to import:
labdata_path <- here::here("data", "Lab_results.xlsx")

# Fuzzy match new lab results to Go.Data cases using names & date of birth:
labmatches <- lab2godata_wrapper(url = url,
                                 username = username,
                                 password = password,
                                 outbreak = "active",
                                 reason = "link new",
                                 datequery = "epibuffer",
                                 daterangeformat = "ymd",
                                 epiwindow = 30,
                                 method = "fuzzy",
                                 matchcols = "names & dob",
                                 labdata = mylabdata,
                                 basedatecol = "sample_collection_date",
                                 firstnamecol = "firstname",
                                 lastnamecol = "surname",
                                 dobcol = "birthdate")
}

}
\seealso{
The wrapper uses the following functions from this (\code{godataR}) package:
\itemize{
\item \code{\link[=get_active_outbreak]{get_active_outbreak()}}
\item \code{\link[=get_access_token]{get_access_token()}}
\item \code{\link[=get_date_range]{get_date_range()}}
\item \code{\link[=mongify_date]{mongify_date()}}
\item \code{\link[=get_cases_epiwindow]{get_cases_epiwindow()}}
\item \code{\link[=get_labresults_epiwindow]{get_labresults_epiwindow()}}
\item \code{\link[=match_cases]{match_cases()}}
}

These functions can be tested with example data included in this package:
\itemize{
\item \link{new_lab_results}
\item \link{case_lookup_table}
\item \link{updated_lab_results}
\item \link{lab_lookup_table}
}
}
\author{
Amy Mikhail, \email{amy.mikhail@gmail.com}
}
