% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_lookup_table.R
\docType{data}
\name{lab_lookup_table}
\alias{lab_lookup_table}
\title{Go.Data exported line list of lab results with identifying columns}
\format{
An object of class \code{data.frame} and \code{data.table}
}
\usage{
data(lab_lookup_table)
}
\description{
Go.Data exported line list of lab results with identifying columns
}
\examples{
# Load data
data(lab_lookup_table)

# Look at head of data set:
head(lab_lookup_table)
}
\author{
Amy Mikhail, \email{amy.mikhail@gmail.com}

Example line list of lab results exported from Go.Data with patient and lab
record identifying columns. This data can be used as a reference (look-up)
table with \code{match_cases()} to demonstrate how to link new information with
existing lab results and update the lab records in Go.Data.  Note that this
is not real data; it has been created for the purpose of demonstration only.
}
\keyword{datasets}
