% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_cols_to_df.R
\name{nested_cols_to_df}
\alias{nested_cols_to_df}
\title{Convert Nested Columns to Data Frames}
\usage{
nested_cols_to_df(df)
}
\arguments{
\item{df}{The input data frame that contains nested columns.}
}
\value{
Returns a data frame. All nested columns will be coerced to data frames.
}
\description{
A helper function to go through datasets and convert any nested columns to data frames.
}
\details{
This function works on all versions of Go.Data. There
are two methods for downloading the data:
}
\examples{
\dontrun{
url <- "https://MyGoDataServer.com/"
username <- "myemail@email.com"
password <- "mypassword"
outbreak_id <- "3b5554d7-2c19-41d0-b9af-475ad25a382b"

cases <- get_cases(url=url,
                   username=username,
                   password=password,
                   outbreak_id=outbreak_id)

cases1 <- nested_cols_to_df(cases)
}
}
