% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null2na.R
\name{null2na}
\alias{null2na}
\title{Convert NULL slots in a data.table to NA}
\usage{
null2na(dtcol)
}
\arguments{
\item{dtcol}{vector containing lists with NULL slots to convert}
}
\value{
the vector with NULLs converted to NAs
}
\description{
This function searches for NULL elements in a vector containing lists and
converts them to the appropriate type of NA. It is useful for transforming
columns containing lists of NULL values from Go.Data, as these columns
cannot be un-nested until the NULL values are converted to NA.
}
\examples{
 \dontrun{
 # Loop through a data.table to convert all NULLs to NAs:
 cases[, names(cases) := lapply(.SD, null2na)]
 }
}
\author{
Amy Mikhail, \email{amy.mikhail@gmail.com}
}
