% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runlab2godata.R
\name{runlab2godata}
\alias{runlab2godata}
\title{Function to launch the lab2godata shiny app}
\usage{
runlab2godata()
}
\description{
This function will launch the lab2godata shiny app, which allows users to
upload lab data and select the required parameters for the
\code{lab2godata_wrapper()} function via a user-friendly interface provided by
the app.  In this way, users with no knowledge of R can run the function.
}
\examples{
\dontrun{
Launch the lab2godata app:
runlab2godata()
}

}
\seealso{
See \code{\link[=lab2godata_wrapper]{lab2godata_wrapper()}} for further information about the function.
}
\author{
Amy Mikhail, \email{amy.mikhail@gmail.com}
}
