% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_active_outbreak.R
\name{set_active_outbreak}
\alias{set_active_outbreak}
\title{Change the currently active outbreak}
\usage{
set_active_outbreak(
  url = url,
  username = username,
  password = password,
  outbreak_id = outbreak_id
)
}
\arguments{
\item{url}{Insert the base URL for your instance of Go.Data here. Don't forget the forward slash "/" at end!}

\item{username}{The email address for your Go.Data login.}

\item{password}{The password for your Go.Data login}

\item{outbreak_id}{The id number for the outbreak to set to active.}
}
\value{
Nothing
}
\description{
A function to change the user's active outbreak
on the Go.Data server. This is a housekeeping
function used in many of the other \code{godataR}
functions.
}
\details{
Each Go.Data user can have 1 and only 1 active
outbreak at a given time.
}
\examples{
\dontrun{
url <- "https://MyGoDataServer.com/"
username <- "myemail@email.com"
password <- "mypassword"
outbreak_id <- "3b5554d7-2c19-41d0-b9af-475ad25a382b"

set_active_outbreak(url=url,
                    username=username,
                    password=password,
                    outbreak_id=outbreak_id)
}
}
