% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_categories.R
\name{translate_categories}
\alias{translate_categories}
\title{Translates cateogories with API labels to more readable forms using the
translation specified in the output of \code{get_language_tokens()}.}
\usage{
translate_categories(data, language_tokens)
}
\arguments{
\item{data}{A data frame (or data frame extension)}

\item{language_tokens}{A data frame (or data frame extension) containing
the translations. Output from \code{get_language_tokens()}}
}
\value{
A tibble
}
\description{
Translates cateogories with API labels to more readable forms using the
translation specified in the output of \code{get_language_tokens()}.
}
\examples{
\dontrun{
cases <- get_cases(
  url = url,
  username = username,
  password = password,
  outbreak_id = outbreak_id
)

language_tokens <- get_language_tokens(
  url = url,
  username = username,
  password = password,
  language = "english_us"
)

translate_categories(
  data = cases,
  language_tokens = language_tokens
)
}
}
