% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_categories.R
\name{translate_token}
\alias{translate_token}
\title{Translates tokens given a specified translation.}
\usage{
translate_token(data, language_tokens)
}
\arguments{
\item{data}{A data frame (or data frame extension)}

\item{language_tokens}{A data frame (or data frame extension) containing
the translations. Output from \code{get_language_tokens()}}
}
\value{
A tibble
}
\description{
This function does all the translation for
\code{translate_categories}.
}
\keyword{internal}
