% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updated_lab_results.R
\docType{data}
\name{updated_lab_results}
\alias{updated_lab_results}
\title{Line list of sequencing results to update existing Go.Data lab records}
\format{
An object of class \code{data.frame} and \code{data.table}
}
\usage{
data(updated_lab_results)
}
\description{
Line list of sequencing results to update existing Go.Data lab records
}
\examples{
# Load data
data(updated_lab_results)

# Look at head of data set:
head(updated_lab_results)
}
\author{
Amy Mikhail, \email{amy.mikhail@gmail.com}

Example line list of whole genome sequencing results that need to be
bulk-imported to Go.Data in order to update existing lab records that only
have the primary diagnostic test result. This data can be used as base data
with \code{match_cases()} to demonstrate how to link sequencing data with existing
lab results in Go.Data and update the lab records.  Note that this is not
real data; it has been created for the purpose of demonstration only.
}
\keyword{datasets}
