% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_godata_url.R
\name{check_godata_url}
\alias{check_godata_url}
\title{Check if the provided Go.Data URL is valid}
\usage{
check_godata_url(url, success_code = 200)
}
\arguments{
\item{url}{Insert the base URL for your instance of Go.Data here. Don't
forget the forward slash "/" at end!}

\item{success_code}{A numeric specifying which code is returned by the API
when successfully returning the status code. Default is 200.}
}
\value{
Boolean, where \code{TRUE} indicates a valid URL.
}
\description{
A function to check whether the provided URL
is valid. This is a housekeeping function
used in many of the other \code{godataR} functions.
}
\examples{
\dontrun{
url <- "https://MyGoDataServer.com/"
check_godata_url(url = url)
}
}
