% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_godata_version.R
\name{check_godata_version}
\alias{check_godata_version}
\title{Check if installed verison of Go.Data is 2.38.1 or later}
\usage{
check_godata_version(url = url)
}
\arguments{
\item{url}{Insert the base URL for your instance of Go.Data here. Don't forget the forward slash "/" at end!}
}
\value{
boolean
}
\description{
Check if installed verison of Go.Data is 2.38.1 or later
}
\examples{
\dontrun{
url <- "https://MyGoDataServer.com/"
check_godata_version(url=url)
}
}
