% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_cases.R
\name{clean_cases}
\alias{clean_cases}
\title{Cleans case data}
\usage{
clean_cases(
  cases,
  cases_address_history_clean,
  cases_vacc_history_clean,
  cases_dateranges_history_clean
)
}
\arguments{
\item{cases}{A \code{tibble} containing the case data.}
}
\value{
A \code{tibble} containing the cleaned case data.
}
\description{
Cleans and un-nests case data. Case data is returned by
\code{\link[=get_cases]{get_cases()}}.
}
\examples{
\dontrun{
url <- "https://MyGoDataServer.com/"
username <- "myemail@email.com"
password <- "mypassword"
outbreak_id <- "3b5554d7-2c19-41d0-b9af-475ad25a382b"

cases <- get_cases(
  url = url,
  username = username,
  password = password,
  outbreak_id = outbreak_id
)

# other cleaned data required for `clean_cases()`
cases_vacc_history_clean <- clean_case_vax_history(cases = cases)
cases_address_history_clean <- clean_case_address_history(cases = cases)
cases_dateranges_history_clean <- clean_case_med_history(cases = cases)

cases_clean <- clean_cases(
  cases = cases,
  cases_address_history_clean = cases_address_history_clean,
  cases_vacc_history_clean = cases_vacc_history_clean,
  cases_dateranges_history_clean = cases_dateranges_history_clean
)
}
}
