% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_access_token.R
\name{get_access_token}
\alias{get_access_token}
\title{Get an access oauth access token for Go.Data}
\usage{
get_access_token(url = url, username = username, password = password)
}
\arguments{
\item{url}{Insert the base URL for your instance of Go.Data here. Don't forget the forward slash "/" at end!}

\item{username}{The email address for your Go.Data login.}

\item{password}{The password for your Go.Data login}
}
\value{
string
}
\description{
A function to retrieve an access token
from Go.Data. This is a housekeeping
function used in many of the other
\code{godataR} functions.
}
\examples{
\dontrun{
url <- "https://MyGoDataServer.com/"
username <- "myemail@email.com"
password <- "mypassword"
get_access_token(url = url,
                 username = username,
                 password = password)
}
}
