% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_export_status.R
\name{get_export_status}
\alias{get_export_status}
\title{Check the status of an export request from Go.Data (version 2.38.1 or later)}
\usage{
get_export_status(
  url = url,
  username = username,
  password = password,
  request_id = request_id
)
}
\arguments{
\item{url}{Insert the base URL for your instance of Go.Data here. Don't forget the forward slash "/" at end!}

\item{username}{The email address for your Go.Data login.}

\item{password}{The password for your Go.Data login}

\item{request_id}{The id number for the export request.}
}
\value{
Returns the current status of the export request.
}
\description{
Check the status of an export request from Go.Data (version 2.38.1 or later)
}
\examples{
\dontrun{
url <- "https://MyGoDataServer.com/"
username <- "myemail@email.com"
password <- "mypassword"
outbreak_id <- "3b5554d7-2c19-41d0-b9af-475ad25a382b"

#Submit an export request
export.request <- GET(paste0(url,"api/outbreaks/",outbreak_id,"/cases/export",
                               "?filter=\%7B\%22where\%22\%3A\%7B\%22useDbColumns\%22\%3A\%22true\%22\%2C\%20\%22dontTranslateValues\%22\%3A\%22true\%22\%2C\%20\%22jsonReplaceUndefinedWithNull\%22\%3A\%22true\%22\%20\%7D\%7D",
                               "&access_token=",get_access_token(url=url, username=username, password=password)))
request_id <- export.request \%>\%
     content() \%>\%
     pluck("exportLogId")

#Check the status of the export request

export.request.status <- get_export_status(url=url, username=username, password=password, request_id=request_id)
}
}
