% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_followups2.R
\name{get_followups2}
\alias{get_followups2}
\title{Download follow-up of contacts from Go.Data (version 2.38.1 or later)}
\usage{
get_followups2(
  url = url,
  username = username,
  password = password,
  outbreak_id = outbreak_id,
  wait = 5,
  file.type = c("json", "csv")
)
}
\arguments{
\item{url}{Insert the base URL for your instance of Go.Data here. Don't forget the forward slash "/" at end!}

\item{username}{The email address for your Go.Data login.}

\item{password}{The password for your Go.Data login}

\item{outbreak_id}{The id number for the outbreak for which you want to download follow-ups.}

\item{wait}{The number of seconds to wait in between iterations of checking the status of the download. Default is 5 seconds, but the user can specify a smaller value to speed up the process if the dataset is small.}

\item{file.type}{Whether the API should return a data structure with nested fields (json, the default) or an entirely flat data structure (csv)}
}
\value{
Returns data frame of follow-ups. Some fields, such as addresses, hospitalization history, and questionnaire fields will require further unnesting. See the tidyr::unnest() function.
}
\description{
Download follow-up of contacts from Go.Data (version 2.38.1 or later)
}
\examples{
\dontrun{
url <- "https://MyGoDataServer.com/"
username <- "myemail@email.com"
password <- "mypassword"
outbreak_id <- "3b5554d7-2c19-41d0-b9af-475ad25a382b"

followups <- get_followups2(url=url, username=username, password=password, outbreak_id=outbreak_id)
}
}
