% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_godata_version.R
\name{get_godata_version}
\alias{get_godata_version}
\title{Quickly check which verison of Go.Data is currently installed.}
\usage{
get_godata_version(url = url)
}
\arguments{
\item{url}{Insert the base URL for your instance of Go.Data here. Don't
forget the forward slash "/" at end!}
}
\value{
string
}
\description{
A function to retrieve the current version
of Go.Data on the provided URL. This is a
housekeeping function used in many of the
other \code{godataR} functions.
}
\examples{
\dontrun{
url <- "https://MyGoDataServer.com/"
get_godata_version(url = url)
}
}
