% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_teams.R
\name{get_teams}
\alias{get_teams}
\title{#' Get teams data from Go.Data}
\usage{
get_teams(url = url, username = username, password = password)
}
\arguments{
\item{url}{Insert the base URL for your instance of Go.Data here. Don't forget the forward slash "/" at end!}

\item{username}{The email address for your Go.Data login.}

\item{password}{The password for your Go.Data login}
}
\value{
Returns a data frame of teams associated with Go.Data instance.
}
\description{
A function to retrieve a list of all teams
data and their attributes, across all outbreaks
on the instance (since this is a system-level API
endpoint). This function relies on the
\verb{\\teams} API endpoint.
}
\examples{
\dontrun{
url <- "https://MyGoDataServer.com/"
username <- "myemail@email.com"
password <- "mypassword"

teams <- get_teams(url=url,
                   username=username,
                   password=password)
}
}
