% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_users.R
\name{get_users}
\alias{get_users}
\title{Get a list of all users and their attributes, across all outbreaks on the instance (since this is a system-level API endpoint).}
\usage{
get_users(url = url, username = username, password = password)
}
\arguments{
\item{url}{Insert the base URL for your instance of Go.Data here. Don't forget the forward slash "/" at end!}

\item{username}{The email address for your Go.Data login.}

\item{password}{The password for your Go.Data login}
}
\value{
Returns data frame of user accounts associated with Go.Data instance. Some fields, roleIds, will be in list format and may require further un-nesting. See the tidyr::unnest() function.
}
\description{
Get a list of all users and their attributes, across all outbreaks on the instance (since this is a system-level API endpoint).
}
\examples{
url <- "https://MyGoDataServer.com/"
username <- "myemail@email.com"
password <- "mypassword"

users <- get_users(url=url, username=username, password=password)
}
